chore: add contactValidations and localStorageKey props to useForm
